requirements.txt
robotframework>=4.1.2
jmespath>=1.0.1
python-dateutil>=2.9.0
requests>=2.31.0
thefuzz>=0.20.0
pyyaml>=6.0.1
jinja2>=3.1.4
tabulate>=0.9.0
google-auth>=2.0.0
google-cloud-bigquery>=3.0.0
docker
azure-containerregistry 
azure-identity